Remove string caching
As much as this seems useful, there are so many other places where Embind is inefficient, and at the same time so much USB work, that this barely makes a difference to speed. It does hurt readability though, so it's better to remove it until we get a better binding system.